refactor(ball_detection): move staged_sampler into data/components
Relocate staged_sampler.py under data/components/ and update the
imports in staged_datamodule.py and its unit test accordingly.
