+++ src/tasks/ball_detection/data/components/staged_sampler.py
